// nivofunnelplus/packages/funnel/src/types.ts
export interface FunnelDatum {
    id: string | number
    value: number
    label?: string
    dataset?: string
}


export interface Position {
    x: number
    y: number
}

export interface BoxPosition extends Position {
    x0: number
    x1: number
    y0: number
    y1: number
}

export type FunnelDirection = 'horizontal' | 'vertical'

export type FunnelAreaPoint = BoxPosition

export type FunnelAreaGenerator = Area<FunnelAreaPoint>

export type FunnelBorderGenerator = Line<Position | null>

export interface FunnelPart<D extends FunnelDatum> extends BoxPosition {
    data: D
    width: number
    height: number
    color: string
    fillOpacity: number
    borderWidth: number
    borderColor: string
    borderOpacity: number
    labelColor: string
    formattedValue: number | string
    isCurrent: boolean
    points: Position[]
    areaPoints: FunnelAreaPoint[]
    borderPoints: (Position | null)[]
}